feat: share network option and choice
