+++ src/ape/click_ext.py
@@ -1,0 +1,23 @@
+import click
+
+from ape import networks
+
+
+class NetworkChoice(click.Choice):
+    """Wraps ``click.Choice`` to provide network choice defaults for the active project."""
+
+    def __init__(self, case_sensitive=True):
+        super().__init__(list(networks.network_choices), case_sensitive)
+
+    def get_metavar(self, param):
+        return "[ecosystem-name][:[network-name][:[provider-name]]]"
+
+
+network_option = click.option(
+    "--network",
+    type=NetworkChoice(case_sensitive=False),
+    default=networks.default_ecosystem.name,
+    help="Override the default network and provider. (see ``ape networks list`` for options)",
+    show_default=True,
+    show_choices=False,
+)
