+++ src/ape_console/_cli.py
@@ -5,29 +5,13 @@ import IPython  # type: ignore
 
 from ape import networks
 from ape import project as default_project
+from ape.click_ext import network_option
 from ape.version import version as ape_version  # type: ignore
-
-
-class NetworkChoice(click.Choice):
-    """Wraps ``click.Choice`` to provide network choice defaults for the active project."""
-
-    def __init__(self, case_sensitive=True):
-        super().__init__(list(networks.network_choices), case_sensitive)
-
-    def get_metavar(self, param):
-        return "[ecosystem-name][:[network-name][:[provider-name]]]"
 
 
 @click.command(short_help="Load the console", context_settings=dict(ignore_unknown_options=True))
 @click.option("--verbose", is_flag=True, flag_value=True, default=False)
-@click.option(
-    "--network",
-    type=NetworkChoice(case_sensitive=False),
-    default=networks.default_ecosystem.name,
-    help="Override the default network and provider. (see ``ape networks list`` for options)",
-    show_default=True,
-    show_choices=False,
-)
+@network_option
 def cli(verbose, network):
     """
     Opens a console for the local project."""
